refactor

+++ src/Controller/Ui/SourceController.php
@@ -7,8 +7,10 @@ use App\Form\SourceType;
 use App\Repository\SourceRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\ResponseHeaderBag;
 use Symfony\Component\Routing\Attribute\Route;
 
 final class SourceController extends AbstractController
@@ -25,7 +27,7 @@ final class SourceController extends AbstractController
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $source = new Source();
-        $form = $this->createForm(SourceType::class, $source);
+        $form   = $this->createForm(SourceType::class, $source);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -37,7 +39,7 @@ final class SourceController extends AbstractController
 
         return $this->render('ui/source/new.html.twig', [
             'source' => $source,
-            'form' => $form,
+            'form'   => $form,
         ]);
     }
 
@@ -63,18 +65,49 @@ final class SourceController extends AbstractController
 
         return $this->render('ui/source/edit.html.twig', [
             'source' => $source,
-            'form' => $form,
+            'form'   => $form,
         ]);
     }
 
     #[Route('/ui/source/{id}', name: 'ui_source_delete', methods: [Request::METHOD_POST])]
     public function delete(Request $request, Source $source, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$source->getId(), $request->getPayload()->getString('_token'))) {
+        if ($this->isCsrfTokenValid('delete' . $source->getId(), $request->getPayload()->getString('_token'))) {
+            $filePath = $source->getFilepath() . "/" . $source->getFilename();
+
+            if (!file_exists($filePath)) {
+                throw new NotFoundHttpException('File not found');
+            }
+            
+            if (unlink($filePath)) {
+                $this->addFlash('success', 'File was deleted');
+            } else {
+                $this->addFlash('error', 'Cannot delete file');
+            }
+
             $entityManager->remove($source);
             $entityManager->flush();
         }
 
         return $this->redirectToRoute('ui_source_index', [], Response::HTTP_SEE_OTHER);
+    }
+
+    #[Route('/ui/source/download/{id}', name: 'ui_source_download', methods: [Request::METHOD_GET])]
+    public function download(Source $source): BinaryFileResponse
+    {
+        $filePath = $source->getFilepath() . "/" . $source->getFilename();
+
+        if (!file_exists($filePath)) {
+            throw $this->createNotFoundException('File was not found');
+        }
+
+        $response = new BinaryFileResponse($filePath);
+
+        $response->setContentDisposition(
+            ResponseHeaderBag::DISPOSITION_ATTACHMENT,
+            $source->getFilename()
+        );
+
+        return $response;
     }
 }
